+++ ports/esp32s2/boards/espressif_hmi_1/board.h
@@ -1,4 +1,4 @@
-/*
+/* 
  * The MIT License (MIT)
  *
  * Copyright (c) 2020 Ha Thach (tinyusb.org) for Adafruit Industries
@@ -22,8 +22,8 @@
  * THE SOFTWARE.
  */
 
-#ifndef MICRODEV_MICROS2_H_
-#define MICRODEV_MICROS2_H_
+#ifndef ESPRESSIF_HMI_1_H_
+#define ESPRESSIF_HMI_1_H_
 
 //--------------------------------------------------------------------+
 // Button
@@ -32,38 +32,58 @@
 // Enter UF2 mode if GPIO is pressed while 2nd stage bootloader indicator
 // is on e.g RGB = Purple. If it is GPIO0, user should not hold this while
 // reset since that will instead run the 1st stage ROM bootloader
-#define PIN_BUTTON_UF2      0
+#define PIN_BUTTON_UF2        0
+
+// GPIO that implement 1-bit memory with RC components which hold the
+// pin value long enough for double reset detection.
+// #define PIN_DOUBLE_RESET_RC   16
 
 //--------------------------------------------------------------------+
 // LED
 //--------------------------------------------------------------------+
 
 // GPIO connected to Neopixel data
-#define NEOPIXEL_PIN        33
+#define NEOPIXEL_PIN          21
 
 // Brightness percentage from 1 to 255
-#define NEOPIXEL_BRIGHTNESS 0x10
+#define NEOPIXEL_BRIGHTNESS   0x30
 
 // Number of neopixels
-#define NEOPIXEL_NUMBER     1
+#define NEOPIXEL_NUMBER       1
 
-// LED for indicator and writing flash
-// If not defined neopixel will be use for flash writing instead
-#define LED_PIN             21
-#define LED_STATE_ON        1
+//Peripheral power is enabled through I2C connected TCA9554
+#define I2C_MASTER_SCL_IO           39
+#define I2C_MASTER_SDA_IO           40
+#define I2C_MASTER_NUM              0
+#define I2C_MASTER_FREQ_HZ          400000
+#define I2C_MASTER_TX_BUF_DISABLE   0
+#define I2C_MASTER_RX_BUF_DISABLE   0
+#define I2C_MASTER_TIMEOUT_MS       1000
+#define I2C_WAIT                    40      //Timing (in microseconds) for I2C
+
+#define TCA9554_ADDR                    0x20
+#define TCA9554_INPUT_PORT_REG          0x00
+#define TCA9554_OUTPUT_PORT_REG         0x01
+#define TCA9554_POLARITY_INVERSION_REG  0x02
+#define TCA9554_CONFIGURATION_REG       0x03
+#define TCA9554_DEFAULT_CONFIG          0b10100000            
+#define TCA9554_DEFAULT_VALUE           0b11100000             //Enable peripheral power and ws2812 data in
+#define TCA9554_PERI_POWER_ON_VALUE     0b11100000             //Enable peripheral power and ws2812 data in
+#define TCA9554_PERI_POWER_OFF_VALUE    0b11110000             //Disable Peripheral power
 
 //--------------------------------------------------------------------+
 // USB UF2
 //--------------------------------------------------------------------+
 
-#define USB_VID             0x239A
-#define USB_PID             0x00C5
-#define USB_MANUFACTURER    "MicroDev"
-#define USB_PRODUCT         "microS2"
+#define USB_VID           0x303A
+#define USB_PID           0x7000
+#define USB_MANUFACTURER  "Espressif"
+#define USB_PRODUCT       "HMI 1"
 
-#define UF2_PRODUCT_NAME    USB_MANUFACTURER " " USB_PRODUCT
-#define UF2_BOARD_ID        "ESP32S2-MICROS2-V1"
-#define UF2_VOLUME_LABEL    "MICROS2BOOT"
-#define UF2_INDEX_URL       "https://github.com/microDev1/microS2/wiki"
+#define UF2_PRODUCT_NAME  USB_MANUFACTURER " " USB_PRODUCT
+#define UF2_BOARD_ID      "ESP32S2-HMI-v1.1"
+#define UF2_VOLUME_LABEL  "ESPHMI1BOOT"
+#define UF2_INDEX_URL     "https://docs.espressif.com/projects/esp-idf/en/latest/esp32s2/hw-reference/esp32s2/user-guide-esp32-s2-kaluga-1-kit.html"
 
-#endif //MICRODEV_MICROS2_H_
+
+#endif