+++ ports/esp32s2/boards/lolin_s2_mini/board.h
@@ -1,4 +1,4 @@
-/*
+/* 
  * The MIT License (MIT)
  *
  * Copyright (c) 2020 Ha Thach (tinyusb.org) for Adafruit Industries
@@ -22,8 +22,7 @@
  * THE SOFTWARE.
  */
 
-#ifndef MICRODEV_MICROS2_H_
-#define MICRODEV_MICROS2_H_
+#pragma once
 
 //--------------------------------------------------------------------+
 // Button
@@ -32,38 +31,43 @@
 // Enter UF2 mode if GPIO is pressed while 2nd stage bootloader indicator
 // is on e.g RGB = Purple. If it is GPIO0, user should not hold this while
 // reset since that will instead run the 1st stage ROM bootloader
-#define PIN_BUTTON_UF2      0
+#define PIN_BUTTON_UF2        0
+
+// GPIO that implement 1-bit memory with RC components which hold the
+// pin value long enough for double reset detection.
+// #define PIN_DOUBLE_RESET_RC
 
 //--------------------------------------------------------------------+
 // LED
 //--------------------------------------------------------------------+
 
-// GPIO connected to Neopixel data
-#define NEOPIXEL_PIN        33
-
-// Brightness percentage from 1 to 255
-#define NEOPIXEL_BRIGHTNESS 0x10
-
-// Number of neopixels
-#define NEOPIXEL_NUMBER     1
-
 // LED for indicator and writing flash
-// If not defined neopixel will be use for flash writing instead
-#define LED_PIN             21
-#define LED_STATE_ON        1
+#define LED_PIN               15
+#define LED_STATE_ON          1
+
+// Uncomment & wire-up NeoPixel to board for testing
+// // GPIOs connected to NeoPixel data & power
+// #define NEOPIXEL_PIN          01
+// 
+// #define NEOPIXEL_POWER_PIN    02
+// #define NEOPIXEL_POWER_STATE  1
+// 
+// // Brightness percentage from 1 to 255
+// #define NEOPIXEL_BRIGHTNESS   0x64
+// 
+// // Number of neopixels
+// #define NEOPIXEL_NUMBER       1
 
 //--------------------------------------------------------------------+
 // USB UF2
 //--------------------------------------------------------------------+
 
-#define USB_VID             0x239A
-#define USB_PID             0x00C5
-#define USB_MANUFACTURER    "MicroDev"
-#define USB_PRODUCT         "microS2"
+#define USB_VID           0x303A
+#define USB_PID           0x80C4
+#define USB_MANUFACTURER  "Lolin"
+#define USB_PRODUCT       "S2 Mini"
 
-#define UF2_PRODUCT_NAME    USB_MANUFACTURER " " USB_PRODUCT
-#define UF2_BOARD_ID        "ESP32S2-MICROS2-V1"
-#define UF2_VOLUME_LABEL    "MICROS2BOOT"
-#define UF2_INDEX_URL       "https://github.com/microDev1/microS2/wiki"
-
-#endif //MICRODEV_MICROS2_H_
+#define UF2_PRODUCT_NAME  USB_MANUFACTURER " " USB_PRODUCT
+#define UF2_BOARD_ID      "ESP32S2FN4R2-S2Mini-1-0-0"
+#define UF2_VOLUME_LABEL  "S2MINIBOOT"
+#define UF2_INDEX_URL     "https://circuitpython.org/board/lolin_s2_mini/"